Add to cart/ Remove from cart refactor

+++ src/app/product/cart-products/cart-products.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from '../../shared/models/product';
 import { ProductService } from '../../shared/services/product.service';
+import { OrderService } from '../../shared/services/order.service';
 @Component({
   selector: 'app-cart-products',
   templateUrl: './cart-products.component.html',
@@ -14,14 +15,30 @@ export class CartProductsComponent implements OnInit {
   messageTitle = 'No Products Found in Cart';
   messageDescription = 'Please, Add Products to Cart';
 
-  constructor(private productService: ProductService) {}
+  constructor(private productService: ProductService,
+    private orderSvc: OrderService) { }
 
   ngOnInit() {
     this.getCartProduct();
   }
 
   removeCartProduct(product: Product) {
-    this.productService.removeLocalCartProduct(product);
+    const userID = localStorage.getItem('userID');
+    const productIds: Number[] = [];
+    productIds.push(product.productId);
+    if (userID === null || userID === undefined) {
+      this.productService.removeLocalCartProduct(product);
+    } else {
+      this.orderSvc.removeFromCart(userID, productIds)
+        .subscribe((data) => {
+          if (data !== null && data !== undefined) {
+            this.productService.removeLocalCartProduct(product);
+          }
+        });
+    }
+
+    // this.productService.removeLocalCartProduct(product);
+
 
     // Recalling
     this.getCartProduct();
